experiment: add side-effect free entrypoints for Button
