+++ dev/register.html
@@ -1,0 +1,22 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Register</title>
+    <script type="module" src="./common.js"></script>
+  </head>
+
+  <body>
+    <vaadin-button>Button</vaadin-button>
+
+    <script type="module">
+      import { Button } from '@vaadin/button/vaadin-button-component.js';
+
+      setTimeout(() => {
+        Button.register();
+      }, 1000);
+    </script>
+  </body>
+</html>
