+++ packages/button/src/vaadin-lit-button-component.d.ts
@@ -3,8 +3,4 @@
  * Copyright (c) 2017 - 2024 Vaadin Ltd.
  * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
  */
-import { Button } from './vaadin-lit-button-component.js';
-
-Button.register();
-
-export { Button };
+export * from './vaadin-button-component.js';
