+++ packages/button/src/vaadin-lit-button-component.js
@@ -1,0 +1,60 @@
+/**
+ * @license
+ * Copyright (c) 2017 - 2024 Vaadin Ltd.
+ * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
+ */
+import { html, LitElement } from 'lit';
+import { defineCustomElement } from '@vaadin/component-base/src/define.js';
+import { ElementMixin } from '@vaadin/component-base/src/element-mixin.js';
+import { PolylitMixin } from '@vaadin/component-base/src/polylit-mixin.js';
+import { TooltipController } from '@vaadin/component-base/src/tooltip-controller.js';
+import { ThemableMixin } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
+import { buttonStyles, buttonTemplate } from './vaadin-button-base.js';
+import { ButtonMixin } from './vaadin-button-mixin.js';
+
+/**
+ * LitElement based version of `<vaadin-button>` web component.
+ *
+ * ## Disclaimer
+ *
+ * This component is an experiment not intended for publishing to npm.
+ * There is no ETA regarding specific Vaadin version where it'll land.
+ * Feel free to try this code in your apps as per Apache 2.0 license.
+ */
+class Button extends ButtonMixin(ElementMixin(ThemableMixin(PolylitMixin(LitElement)))) {
+  static get is() {
+    return 'vaadin-button';
+  }
+
+  static get styles() {
+    return buttonStyles;
+  }
+
+  /** @protected */
+  static registerStyles() {
+    // To be overridden by the theme
+  }
+
+  static register() {
+    if (!customElements.get(this.is)) {
+      this.registerStyles();
+
+      defineCustomElement(this);
+    }
+  }
+
+  /** @protected */
+  render() {
+    return buttonTemplate(html);
+  }
+
+  /** @protected */
+  ready() {
+    super.ready();
+
+    this._tooltipController = new TooltipController(this);
+    this.addController(this._tooltipController);
+  }
+}
+
+export { Button };
