+++ packages/button/theme/material/vaadin-button.js
@@ -1,2 +1,5 @@
-import './vaadin-button-styles.js';
-import '../../src/vaadin-button.js';
+import { Button } from './vaadin-button-component.js';
+
+Button.register();
+
+export { Button };
