+++ packages/button/vaadin-button-component.js
@@ -1,0 +1,1 @@
+export * from './theme/lumo/vaadin-button-component.js';
