+++ packages/button/vaadin-button.js
@@ -1,3 +1,3 @@
 import './theme/lumo/vaadin-button.js';
 
-export * from './src/vaadin-button.js';
+export * from './theme/lumo/vaadin-button.js';
