+++ packages/button/vaadin-lit-button.js
@@ -1,3 +1,3 @@
 import './theme/lumo/vaadin-lit-button.js';
 
-export * from './src/vaadin-lit-button.js';
+export * from './theme/lumo/vaadin-lit-button.js';
