+++ packages/date-picker/theme/lumo/vaadin-date-picker.js
@@ -1,3 +1,4 @@
+import '@vaadin/button/theme/lumo/vaadin-button.js';
 import './vaadin-date-picker-overlay-styles.js';
 import './vaadin-date-picker-overlay-content-styles.js';
 import './vaadin-month-calendar-styles.js';
