+++ packages/rich-text-editor/theme/material/vaadin-lit-rich-text-editor.js
@@ -1,4 +1,4 @@
-import '@vaadin/button/theme/material/vaadin-button-styles.js';
+import '@vaadin/button/theme/material/vaadin-lit-button.js';
 import '@vaadin/confirm-dialog/theme/material/vaadin-confirm-dialog-styles.js';
 import '@vaadin/text-field/theme/material/vaadin-text-field-styles.js';
 import '@vaadin/tooltip/theme/material/vaadin-tooltip-styles.js';
